requirements.txt
backports.statistics==0.1.0
Cerberus==1.3.4
ConfigParser==6.0.0
Cython==3.0.0
dl==0.1.0
ipaddr==2.2.0
jsonschema==3.2.0
keyring==24.2.0
mypy_extensions==1.0.0
mysql_connector_repackaged==0.3.1
numpy==1.24.4
ordereddict==1.1
pandas==2.0.3
Requests==2.31.0
wincertstore==0.2
mysql-connector-python==8.1.0